# Add support for BSWAP intrinsic