--- a/src/jit/compiler.h
+++ b/src/jit/compiler.h
@@ -9741,6 +9741,8 @@ public:
             // Standard unary operators
             case GT_NOT:
             case GT_NEG:
+            case GT_BSWAP:
+            case GT_BSWAP16:
             case GT_COPY:
             case GT_RELOAD:
             case GT_ARR_LENGTH: