--- a/src/jit/namedintrinsiclist.h
+++ b/src/jit/namedintrinsiclist.h
@@ -9,11 +9,12 @@
 
 enum NamedIntrinsic : unsigned short
 {
-    NI_Illegal                                                 = 0,
-    NI_System_Enum_HasFlag                                     = 1,
-    NI_MathF_Round                                             = 2,
-    NI_Math_Round                                              = 3,
-    NI_System_Collections_Generic_EqualityComparer_get_Default = 4,
+    NI_Illegal                                                  = 0,
+    NI_System_Enum_HasFlag                                      = 1,
+    NI_MathF_Round                                              = 2,
+    NI_Math_Round                                               = 3,
+    NI_System_Collections_Generic_EqualityComparer_get_Default  = 4,
+    NI_System_Buffers_Binary_BinaryPrimitives_ReverseEndianness = 5,
 #ifdef FEATURE_HW_INTRINSICS
     NI_HW_INTRINSIC_START,
 #if defined(_TARGET_XARCH_)