Addresses https://github.com/dotnet/coreclr/issues/14122

The general idea is that we expose APIs (for internal consumption) which are JIT intrinsics, replaced with `BSWAP` instructions instead of method calls at runtime. I've put them on `BitConverter` for now but haven't exposed them via the reference API. The System.Memory package already exposes `BinaryPrimitives.ReverseEndianness` APIs, and those can call through to these implementations due to the special relationship between corefx packages and System.Private.CoreLib.

Here's the x64 codegen of several `test_bswap` routines that I compiled and tested.

```asm
test_bswap(UInt32):
00007ffd`ac470080 8bc1            mov     eax,ecx
00007ffd`ac470082 0fc8            bswap   eax
00007ffd`ac470084 c3              ret

test_bswap(UInt64):
00007ffd`ab6a0080 488bc1          mov     rax,rcx
00007ffd`ab6a0083 480fc8          bswap   rax
00007ffd`ab6a0086 c3              ret

test_bswap(UInt16):
00007ffd`ab6d0080 0fb7c1          movzx   eax,cx
00007ffd`ab6d0083 66c1c808        ror     ax,8
00007ffd`ab6d0087 c3              ret
```

Open issues:

* How would I even begin to test this in a sane fashion? Right now I tested this by using a `DynamicMethod` to reference these intrinsic APIs, then compiling and observing the codegen in windbg. It works as far as seeing what's going on, but I can't use this for things like perf runs.

* An alternative implementation would be to recognize byte swap IL patterns directly, similar to how rotation IL patterns are recognized directly. I didn't pursue this because I thought it would be more error-prone than using special intrinsic methods, but perhaps somebody knows a more foolproof way of doing this.

* I didn't attempt to make this work on ARM because I don't have a test box. The ARM equivalents are `REV` / `REV16` / `REV64`. I may need to suppress the intrinsic entirely on ARM for now in order to avoid JIT breaks.

* Calling `ReverseEndianness(<some const uint64>)` works but isn't optimized properly. I think I need to update morph but I'll need assistance knowing where to look.

* Ideally if we detect a `GT_BSWAP` instruction immediately before or after a memory access, we can replace both operations with `movbe` on compatible architectures. I don't think this would have much impact on runtime, but it could save a scratch register and help avoid stack spillage in some scenarios.