@GrabYourPitchforks 2018-06-09
I managed to get some perf numbers from before and after the change. The baselines here are the existing implementations of `BinaryPrimitives.ReverseEndianness`. Measurements are total time taken to call the method 1 billion times in a loop. I confirmed via code gen inspection that no intermediate calculations were being discarded.

**Before `bswap` enablement**

* ushort: 915 ms
* uint: 1120 ms
* ulong: 2120 ms

**After `bswap` enablement**

* ushort: 330 ms
* uint: 400 ms
* ulong: 590 ms