@jkotas 2018-06-10
> System.Memory package already exposes BinaryPrimitives.ReverseEndianness APIs, and those can call through to these implementation

We have tried similar hops in the past (e.g. with Span), but they never survived for long. It may be better to recognize `BinaryPrimitives.ReverseEndianness` as intrinsic directly - perhaps move `BinaryPrimitives` to CoreLib.