@GrabYourPitchforks 2018-07-12
@dotnet-bot test Windows_NT x64 Formatting