@GrabYourPitchforks 2018-07-12
Rebased and responded to some PR comments. This should address the majority of the reliability feedback that reviews have given (thanks all!), but it's still WIP.

**Summary of changes:**

* Condition the bswap intrinsic logic on a check for x86 / x64
* Removed `GT_BSWAP` logic from `genGetInsForOper` and moved it all into `genCodeForBswap`
* Use `GT_BSWAP16` instead of `GT_BSWAP` for 16-bit byte swap operations
* Enlighten morph's const folding to understand `GT_BSWAP` / `GT_BSWAP16`

This fixed a handful of problems the original prototype had, including mishandling of 64-bit const values. I tested both consts and non-consts across all three overloads on both x86 and x64, and the behavior was correct in all cases. If a const is provided, regardless of arch, constant folding will take over. If a non-const is provided, it emits a `ror reg.16, 8` or `bswap reg` instruction as appropriate, except for the 64-bit overload on 32-bit archs, which gets decomposed into two `bswap reg` instructions.

**To-do:**

* Not expose these methods on `BitConverter`; instead consider moving `BinaryPrimitives` into corelib (needs API review?)
* Add tests (how?)
* Add support for movbe (probably not as part of this PR? significantly impacts AoT compilation.)