@jkotas 2018-07-13
> Not expose these methods on BitConverter; instead consider moving BinaryPrimitives into corelib (needs API review?)

Moving implementations of existing APIs to CoreLib does not need API review. It should be done in separate PR that just moves the implementation, and does not have any other changes.