@GrabYourPitchforks 2018-10-09
I believe all open issues are resolved as of the latest iteration of the PR. Here's the current status:

* This causes the JIT to special-case `BinaryPrimitives.ReverseEndianness` instead of introducing a new internal method `BitConverter.ByteSwap`.
* On 32-bit x86, the 16-bit and 32-bit overloads of `ReverseEndianness` are replaced with intrinsics. The 64-bit overload calls into the 32-bit overloads as helper methods.
* On x64, the 16-bit, 32-bit, and 64-bit overloads of `ReverseEndianness` are treated as intrinsics.
* On arm, none of the overloads are treated as intrinsics, and the runtime falls back to the existing software implementation.
* I verified that constant folding is handled as expected by the changes in gentree.cpp.
* There are tests both here and in corefx verifying the correct behavior of these APIs.

The only thing I wasn't able to get to was providing the 64-bit fallback implementation entirely in the VM. If this is still important, I hope it's acceptable to allow this PR to go through as-is (modulo any new feedback) and to open a new work item for providing the VM fallback.

Removing WIP from title.