@GrabYourPitchforks 2018-10-10
I rebased the PR since there were some recent commits that addressed the flaky unit tests.