@GrabYourPitchforks 2018-10-11
@dotnet-bot test OSX10.12 x64 Checked Innerloop Build and Test