@GrabYourPitchforks 2018-11-05
@dotnet/jit-contrib Any final feedback on this? I have an open question re: `Compiler::fgMorphSmpOp` but it doesn't seem to affect correctness.