@mikedn 2018-11-05
Sorry, I missed your earlier question. I don't think that the `fgMorphSmpOp` is really useful. It's not like having a constant op node is breaking some kind of rule. It's surprising that a noway_assert is used there.